Downgrade wkhtmltopdf
The latest wkhtml version provided by Debian is incompatible with odoo and probably won't be maintained anymore, the reason is that Odoo os changing the ORM for other ways to manipulated PDFs. Te monkeypatch i found is to use the specific use of this: https://github.com/odoo/odoo/wiki/Wkhtmltopdf specially for Odoo16.0 https://github.com/wkhtmltopdf/packaging/releases/tag/0.12.6.1-3
The problem here is that it is CPU oriented. The main usage here for me has been AMD64, so probably you should also check this for your self